LogFactory - Disconnect from Target write and Target flush
